+++ Cargo.toml
@@ -27,10 +27,9 @@ integer-sqrt = "0.1.5"
 jsonrpc = "0.12.1"
 lazy_static = "1.4.0"
 log = "0.4.14"
-ord-lmdb-zero = "0.4.5"
 qrcode-generator = "4.1.6"
 rayon = "1.5.1"
-redb = { version = "0.0.5", optional = true }
+redb =  "0.3.0"
 reqwest = { version = "0.11.10", features = ["blocking", "json"] }
 secp256k1 = { version =  "0.22.1", features = ["rand", "rand-std", "global-context"] }
 serde = { version = "1.0.137", features = ["derive"] }
